make jailed as a es6 module and move it to src; add tests
