+++ web/public/index.html
@@ -81,7 +81,7 @@
     <!-- <script src="/static/js/filtrex.js" type="text/javascript"></script> -->
     <!-- <link rel="stylesheet" href="/static/leaflet/1.3.1/leaflet.css"/> -->
     <!-- <script src="/static/leaflet/1.3.1/leaflet.js"></script> -->
-    <script src="/static/jailed/jailed.js"></script>
+    <!-- <script src="/static/jailed/jailed.js"></script> -->
     <!-- <script src="/static/js/pouchdb-7.0.0.min.js"></script> -->
     <style>
     @font-face {
